Time: 0 ms (100%), Space: 23.1 MB (67.03%) - LeetHub

+++ 3242-count-elements-with-maximum-frequency/3242-count-elements-with-maximum-frequency.cpp
@@ -1,8 +1,8 @@
 class Solution {
 public:
     int maxFrequencyElements(vector<int>& nums) {
-        uint8_t freq[101] = {0};
-        uint8_t max = 0, res = 0;
+        int freq[101] = {0};
+        int max = 0, res = 0;
         for (int n : nums) {
             uint8_t f = ++freq[n];
             if (f > max) {
